fix(A.6): standardize double precision in domain financial models

+++ app/src/main/java/com/yourname/expensetracker/data/repository/DashboardContractsAdapter.kt
@@ -71,7 +71,7 @@ class DashboardContractsAdapter @Inject constructor(
                     categoryName = status.category?.name,
                     spentAmount = status.spentAmount,
                     remainingAmount = status.remainingAmount,
-                    percentUsed = status.percentUsed,
+                    percentUsed = status.percentUsed.toDouble(),
                     healthStatus = status.healthStatus,
                     periodStart = status.periodStart,
                     periodEnd = status.periodEnd
