+++ app/src/main/java/com/yourname/expensetracker/data/repository/FinancialWeatherRepository.kt
@@ -73,7 +73,7 @@ class FinancialWeatherRepository @Inject constructor(
                 categoryName = status.category?.name,
                 spentAmount = status.spentAmount,
                 remainingAmount = status.remainingAmount,
-                percentUsed = status.percentUsed,
+                percentUsed = status.percentUsed.toDouble(),
                 healthStatus = status.healthStatus,
                 periodStart = status.periodStart,
                 periodEnd = status.periodEnd
